requirements.txt
anyio==3.7.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
camelot-py==0.11.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.95.2
ghostscript==0.7
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
numpy==1.21.6
opencv-python==4.7.0.72
openpyxl==3.1.2
outcome==1.2.0
pandas==1.3.5
pdfminer.six==20221105
pycparser==2.21
pydantic==1.10.8
pypdf==3.9.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
tabula-py==2.6.0
tabulate==0.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
urllib3==2.0.2
wsproto==1.2.0
XlsxWriter==3.1.2
zipp==3.15.0
